Extract tab/pane lifecycle into WindowTabsMixin
Continue decomposing window.py with the largest cluster: move the contiguous
44-method tab/pane block — tab close/rename/title, the tab context menus and
their _on_tabmenu_* handlers, file-manager embed teardown (shared with
shutdown), tab attach/detach, drag-to-convert-to-split, and the layout toggle —
verbatim into sshpilot/window_tabs.py as WindowTabsMixin, inherited by
MainWindow. window.py shrinks ~1066 lines (now ~8.7k).

Pure code move. SplitViewTab stays imported locally inside the methods that use
it (a module-level import would be circular: split_view -> window ->
window_tabs); the other module-level names the block relied on
(TerminalWidget, icon_utils, Capability/capabilities_for, connection
host/alias, launch_remote_file_manager, the preferences visibility helpers) are
imported into the mixin so behavior is unchanged.

Verified no moved name remains in the window.py body (no MRO shadow), and
MainWindow has no Gtk.Template so the relocated signal handlers stay wired via
explicit self.connect.

Add tests/test_window_tabs_mixin.py asserting all 44 moved methods resolve to
the mixin module and the mixin is in the MRO.
